requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# Utils
django-debug-toolbar==4.2.0
django-browser-reload==1.12.1
Pillow==10.0.1

# Services
celery==5.3.4
redis==5.0.1
hiredis==2.2.3

# Tooling
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

# Authentication
social-auth-app-django


# tracking
django-user-visit==2.0

# DB Layer
# psycopg2-binary==3.1.12
# mysqlclient==2.2.0

# Species Classification
# pytorch==1.2.0
# torchvision==0.4.0
# cudatoolkit==10.0
# pretrainedmodels==0.7.4
# pillow==6.1.0
# progressbar2==3.51.0
# cupy-cuda100==7.3.0
# torchnet==0.0.4